Moved commands, message and messages to directory
