+++ lib/index.js
@@ -2,15 +2,10 @@
  * @namespace P2P
  */
 
-var Messages = require('./messages');
-
 module.exports = {
-  Message: require('./message'),
-  Commands: require('./commands'),
   Inventory: require('./inventory'),
   BloomFilter: require('./bloomfilter'),
-  Messages: Messages,
-  messages: new Messages(),
+  Messages: require('./messages'),
   Peer: require('./peer'),
   Pool: require('./pool')
 };
