+++ lib/messages/commands.js
@@ -1,14 +1,14 @@
 'use strict';
 
 var Message = require('./message');
+var BloomFilter = require('../bloomfilter');
+var packageInfo = require('../../package.json');
 var inherits = require('util').inherits;
-var packageInfo = require('../package.json');
 var bitcore = require('bitcore');
 var BN = bitcore.crypto.BN;
 var BufferReader = bitcore.encoding.BufferReader;
 var BufferWriter = bitcore.encoding.BufferWriter;
 var BufferUtil = bitcore.util.buffer;
-var BloomFilter = require('./bloomfilter');
 var Put = require('bufferput'); //todo remove
 var $ = bitcore.util.preconditions;
 var _ = bitcore.deps._;
