+++ lib/messages/index.js
@@ -3,10 +3,9 @@
 var bitcore = require('bitcore');
 var BufferUtil = bitcore.util.buffer;
 var Hash = bitcore.crypto.Hash;
-var Commands = require('./commands');
 
 function Messages(options) {
-  this.commands = new Commands(options);
+  this.commands = new Messages.Commands(options);
 
   if (!options) {
     options = {};
@@ -23,6 +22,8 @@ function Messages(options) {
 
 Messages.MINIMUM_LENGTH = 20;
 Messages.PAYLOAD_START = 16;
+Messages.Message = require('./message');
+Messages.Commands = require('./commands');
 
 Messages.prototype.parseMessage = function(dataBuffer) {
   /* jshint maxstatements: 18 */
