+++ lib/messages/message.js
