+++ lib/pool.js
@@ -38,6 +38,8 @@ function now() {
  * @constructor
  */
 function Pool(options) {
+  /* jshint maxcomplexity: 10 */
+  /* jshint maxstatements: 20 */
 
   var self = this;
